+++ src/systems/eu/implementations/react/components/list/examples/Description.jsx
@@ -1,0 +1,6 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import React from 'react';
+import demoDescription from '@ecl/eu-specs-list/demo/data--description';
+import DescriptionListWithData from '../src/DescriptionListWithData';
+
+export default () => <DescriptionListWithData {...demoDescription} />;
