+++ src/systems/eu/implementations/react/components/list/examples/Ordered.jsx
@@ -1,0 +1,6 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import React from 'react';
+import demoText from '@ecl/eu-specs-list/demo/data--text';
+import OrderedListWithData from '../src/OrderedListWithData';
+
+export default () => <OrderedListWithData {...demoText} />;
