+++ src/website/package.json
@@ -71,7 +71,6 @@
     "serve": "11.0.0",
     "style-loader": "0.23.1",
     "terser-webpack-plugin": "1.2.3",
-    "unescape-es6": "2.1.2",
     "url-loader": "1.1.2",
     "webpack": "4.31.0",
     "webpack-bundle-analyzer": "3.3.2",
